requirements.txt
asgiref==3.4.1
astroid==2.2.5
certifi==2021.5.30
charset-normalizer==2.0.5
cloudinary==1.26.0
colorama==0.4.1
Django==3.2.7
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-js-asset==1.2.2
idna==3.2
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==8.3.2
psycopg2==2.9.1
pylint==2.3.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typed-ast==1.4.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wincertstore==0.2
wrapt==1.11.2
